pyproject.toml
[project]
name = "enrollmentapp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "streamlit",
    "google-genai",
    "python-docx",
    "openpyxl",
    "python-pptx"
]
